+++ docs_manual/source/conf.py
@@ -26,7 +26,7 @@ copyright = '2018-2025, BIH Core Unit Bioinformatics'
 author = 'BIH Core Unit Bioinformatics'
 
 # The full version, including alpha/beta/rc tags
-release = '1.1.1'
+release = '1.1.2-WIP'
 
 
 # -- General configuration ---------------------------------------------------